Let emoji_store.jsx use Client APIs to resolve URLs
This reverts commit 948b557453550646ad3213cb4144055eb7db0d69 and
introduces a proper fix for testing.

Commit 948b557453550646ad3213cb4144055eb7db0d69 reverted some changes
in commit 7e9cf13aa356f991f48ba0a943bdab9b4d3c9233. Commit
7e9cf13aa356f991f48ba0a943bdab9b4d3c9233 used module aliasing to prevent
Node.js, used for testing, to import browser-dependent modules. The
module aliasing was dependent on a relative path, which is not reliable,
and it actually got broken with commit
177589b1e26fcabd7749dd0fbe8c39999c206609.

This commit includes an alternative fix for the issue; utilization of
the following specification, which allows to alias by specifying
with relative pathes.
https://github.com/defunctzombie/package-browser-field-spec
